refactor(health): split diagnostics checks by concern
